+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-<!-- markdownlint-disable-file MD024 -->
-
 [unreleased]: https://gitlab.com/tozd/dinit/-/compare/v0.1.0...main
 [0.1.0]: https://gitlab.com/tozd/dinit/-/tags/v0.1.0
+
+<!-- markdownlint-disable-file MD024 -->