+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - If `run` file exits with code 115 it signals that the program is disabling itself
   and that it does not have to run and the rest of the container is then not terminated.
-- Change name to `stop` file instead of `finish` file to avoid confusion.
+- Change name to `terminate` file instead of `finish` file to avoid confusion.
   `finish` file is used differently in runit.
 
 ## [0.1.0] - 2023-07-09