+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-07-10
+
 ### Changed
 
 - If `run` file exits with code 115 it signals that the program is disabling itself
@@ -20,7 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[unreleased]: https://gitlab.com/tozd/dinit/-/compare/v0.1.0...main
+[unreleased]: https://gitlab.com/tozd/dinit/-/compare/v0.2.0...main
+[0.2.0]: https://gitlab.com/tozd/dinit/-/compare/v0.1.0...v0.2.0
 [0.1.0]: https://gitlab.com/tozd/dinit/-/tags/v0.1.0
 
 <!-- markdownlint-disable-file MD024 -->